[tiny] serverLoadData method only usable from node imports
