+++ packages/gasket-helper-intl/package.json
@@ -2,7 +2,8 @@
   "name": "@gasket/helper-intl",
   "version": "6.0.0",
   "description": "Internal helpers used by loaders to resolve locale file paths",
-  "main": "lib",
+  "main": "lib/server.js",
+  "browser": "lib/index.js",
   "scripts": {
     "lint": "eslint .",
     "lint:fix": "npm run lint -- --fix",
@@ -12,7 +13,7 @@
     "test:coverage": "nyc --reporter=text --reporter=json-summary npm run test:runner",
     "posttest": "npm run lint",
     "prepack": "npm run docs",
-    "docs": "jsdoc2md --plugin @godaddy/dmd --files lib/*.js > docs/api.md"
+    "docs": "jsdoc2md --plugin @godaddy/dmd --files lib/index.js > docs/api.md"
   },
   "repository": {
     "type": "git",
